+++ src/resources/assets/js/components/labelListLabelItem.vue
